prototype/requirements.txt
requests>=2.31.0
pillow>=11.0.0
numpy>=1.26.0
opencv-python>=4.10.0
python-dotenv>=1.0.0
lottie>=0.1.1
